refactor: move quiz validation to suggested_quizzes helper
